# Full Stack/backend/routes/swagger.yaml
swagger: '2.0'
info:
  description: |
    This is a sample MobileStore server.
  version: 1.0.0
  title: Heiki ja Pavel Mobilestore
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@pavel.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# host: petstore.swagger.io
# basePath: /v2
...
paths:
...
  /mobile/{mobileId}:
    get:
      tags:
      - mobile
      summary: Find mobile by ID
      description: Returns a single mobile
      operationId: getMobileById
      produces:
      - application/json
      - application/xml
      parameters:
      - name: mobileId
        in: path
        description: ID of mobile to return
        required: true
        type: integer
        format: int64
      responses:
        201:
          description: successful operation
          schema:
            $ref: '#/definitions/Mobile'
        400:
          description: Invalid ID supplied
        404:
          description: Mobile not found
      security:
      - mobilestore_auth: []
    post:
      tags:
      - mobile
      summary: Updates a mobile in the store with form data
      operationId: updateMobileWithForm
      consumes:
      - application/x-www-form-urlencoded
      produces:
      - application/json
      - application/xml
      parameters:
      - name: mobileId
        in: path
        description: ID of mobile that needs to be updated
        required: true
        type: integer
        format: int64
      - name: name
        in: formData
        description: Updated name of the mobile
        required: false
        type: string
      - name: status
        in: formData
        description: Updated status of the mobile
        required: false
        type: string
      responses:
        405:
          description: Invalid input
      security:
      - mobilestore_auth: []
    delete:
      tags:
      - mobile
      summary: Deletes a mobile
      operationId: deleteMobile
      produces:
      - application/json
      - application/xml
      parameters:
      - name: api_key
        in: header
        required: false
        type: string
      - name: mobileId
        in: path
        description: Mobile id to delete
        required: true
        type: integer
        format: int64
      responses:
        201:
          description: User deleted
        400:
          description: Invalid ID supplied
        404:
          description: Mobile not found
      security:
      - mobilestore_auth: []
...
securityDefinitions:
  mobilestore_auth:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  Mobile:
    type: object
    required:
    - title
    - description
    - imageUrl
    - userId
    - price
    properties:
      id:
        type: integer
        format: int64
      title:
        type: string
        example: phone
      description:
        type: string
        example: cool new phone
      imageUrl:
        type: string
      userId:
        type: string
      price: 
        type: number
    xml:
      name: Mobile
# Added by API Auto Mocking Plugin
host: localhost:3000
basePath: /api
schemes:
 - http